+++ nimrod/__main__.py
@@ -5,9 +5,9 @@ from nimrod.dynamic_analysis.criteria.first_semantic_conflict_criteria import Fi
 from nimrod.dynamic_analysis.criteria.second_semantic_conflict_criteria import SecondSemanticConflictCriteria
 from nimrod.dynamic_analysis.main import DynamicAnalysis
 from nimrod.core.merge_scenario_under_analysis import MergeScenarioUnderAnalysis
-from nimrod.reports_generation.behavior_change_output_generator import BehaviorChangeOutputGenerator
-from nimrod.reports_generation.semantic_conflicts_output_generator import SemanticConflictsOutputGenerator
-from nimrod.reports_generation.test_suites_output_generator import TestSuitesOutputGenerator
+from nimrod.output_generation.behavior_change_output_generator import BehaviorChangeOutputGenerator
+from nimrod.output_generation.semantic_conflicts_output_generator import SemanticConflictsOutputGenerator
+from nimrod.output_generation.test_suites_output_generator import TestSuitesOutputGenerator
 from nimrod.smat import SMAT
 from nimrod.test_suite_generation.main import TestSuiteGeneration
 from nimrod.tests.utils import setup_logging, get_config
